arch_overlay/qc_iu: Fix bug in qc.shlsat
The shift X[rs1] << X[rs2][4:0] has the result type of
typeof(X[rs1]) which is an XLEN-bit integer and will therefore not
capture overflows.  Instead, use a sign-extended X[rs1] as the LHS
in the shift operation.

Fix parenthesis in calculation of the most positive signed XLEN integer
(<< has lower precedence than -).

Signed-off-by: Anton Johansson <anjo@rev.ng>