Fix RORW operand and IDL
The YAML for the RORW instruction has shamt as an operand. shamt is used as an immediate
value, but the RORW instruction uses a register value. The RISC-V ISA specification shows the
analogous RORW instruction's operand is named rs2, which is appropriate here.

Change RORW operand shamt to rs2.

Also, in the IDL for RORW, the shamt _value_ should be retrieved from rs2 (not rs1,
as is done currently.

Fix the retrieval of the shamt value to come from rs2.