fix: skip fence.tso on llvm-test since compiler and ISA view of instruction differ
fence.tso is a fence instruction and the ISA states it shall be all zeros. Compilers like LLVM take this literally and only implement it all constant, while the ISA leaves some fields open to newer implementation approaches. When we check UDB version that refers to the ISA vs LLVM, the test fails because of the different view on the subject.
Ad-hoc skip the LLVM test for fence.tso.

Signed-off-by: Afonso Oliveira <Afonso.Oliveira@synopsys.com>