feat(ruby): add static type checking with Sorbet (#585)
Ruby, like most scripting languages, is not statically typed. While
great for initial productivity, it can make changes challenging since
errors are only caught at runtime, and only if that path is exercised.

There are two type checking solutions for Ruby:
[Sorbet](https://sorbet.org/docs/overview) and
[RBS/Steep](https://github.com/soutaro/steep). I tried both with the UDB
code base. Based on that evaluation, I determined that Sorbet was higher
quality, had much better documentation, and more mature. It's also in
commercial use so is likely to be supported long-term.

Sorbet uses _gradual type checking_: you can add type signatures to
subsets of the code, and Sorbet only checks what is known. Thus, we can
start using it without annotating the entire code base up front.

This PR highlights the utility of the tool: it found several bugs
already.

There is a *cost* to using Sorbet, mostly in terms of added development
time

* The signature syntax needs to be understood by developers.
* Some extra [assertions]https://sorbet.org/docs/type-assertions are
needed in code aside from the function signatures
* Though I found the tool is surprisingly helpful in telling you exactly
what assertion is needed (even with an autocorrect option)
 

Let's discuss more here, but based on this experiment my recommendation
is that we merge this PR.