fix: correct "writeable" to "writable" (#728)
Both are used in places. While "writeable" is generally defined as "an
alternative spelling of 'writable'", the latter seems to be generally
accepted as correct/preferred. The RISC-V ISA specification uses
"writable".

BREAKING CHANGE: Change spelling of "writeable" to "writeable"
everywhere

The spelling of "writeable" is corrected to "writable" everywhere,
including CSR schema and configuration parameters.