fix: too much quoting in `lib/deploy.sh`
Commit 4ddb999 added some nice logging, but also added a lot of quotes.
A few too many, in fact, as this:
```
./do gen:html_manual MANUAL_NAME=isa VERSIONS=all
```
became:
```
deploy_do "gen:html_manual MANUAL_NAME=isa VERSIONS=all"
```

The quotes there turn "command argument argument" into a single command
string where the spaces are part of the command itself.

Remove the quotes so that the arguments are again treated as arguments.