C header generation support (#553)
This need to go after the Golang support, I'll fix the commit history
once that is merged.

For the output difference check:
https://www.diffchecker.com/22skd4o8/


There are some differences that relate to how UDB views the data, but I
think that is not entirely related to this PR and needs proper
discussion on UDB vs riscv-opcodes on those particular instructions.

Also, ordering is messed up, but riscv-opcodes does not enforce any
ordering so it is impossible to match the original order

---------

Signed-off-by: Afonso Oliveira <Afonso.Oliveira@synopsys.com>