style: add assembly syntax to generated instruction appendix
It's time.

After a number of recent commits to both UDB and RISC-V Sail repositories,
the only differences between the two with respect to assembly syntax are:
- `amo*.{aq,rl}` missing in UDB
- `mop.r.n`, `mop.rr.n`, `c.mop.n` appear unexpanded
- issue #821
- `vsetvli`, `vsetivli` have 2 forms of assembly syntax, and
  only a single syntax can be expressed currently in UDB

(And there are a number of instructions missing in Sail that are in UDB.)

Make the assembly syntax unconditionally appear in the generated
instruction appendix.