feat(ci): add GitHub merge workflow (#880)
This adds a workflow that should run when a PR is added to the merge
queue. The workflow has all the same jobs that run to deploy pages. This
is step 1 of a full fix; step two is to deploy using the same GitHub
actions instead of the deploy.sh script.