feat: install Python and Node packages in container (#1220)
Python, Node, and Ruby packages were all installed in the local
workspace at runtime. This slows down initial runs and creates the
potential for people to use different versions of packages (the Node
lock file was not actually being used). It also creates permissions
issues when switching between different container runtimes (I've been
switching between Podman and the Devcontainer lately).

This installs the Python and Node packages in the container and updates
the `npm` installation process to actually use `package-lock.json`.

~~I haven't tested with Singularity~~ (I'm a Podman/Docker user), but
everything is passing with Podman locally.

I'll follow up with another PR later to migrate some of the Ruby
packages to the container as well, but that one is trickier because some
of the Ruby packages are local to the UDB project.